refactor: Create recipe function
