+++ getrecipe/services/recipe/main.py
@@ -1,20 +1,22 @@
 from . import scraped_recipe
+from . import scraper
 from .model import ScrapedRecipe
 
 
 def get_scraped_recipe(url):
     try:
-        soup = scraped_recipe.soup(url)
+        soup = scraper.soup(url)
+        recipe = scraped_recipe.recipe(soup)
     except Exception as e:
         print(f"Error: {e}")
         return None
 
     return ScrapedRecipe(
         url=url,
-        title=scraped_recipe.title(soup),
-        ingredients=scraped_recipe.ingredients(soup),
-        instructions=scraped_recipe.instructions(soup),
-        servings=scraped_recipe.servings(soup),
-        prep_time=scraped_recipe.preptime(soup),
-        cook_time=scraped_recipe.cooktime(soup),
+        title=recipe["title"],
+        ingredients=recipe["ingredients"],
+        instructions=recipe["instructions"],
+        servings=recipe["servings"],
+        prep_time=recipe["preptime"],
+        cook_time=recipe["cooktime"],
     )
